refactor: extract window policy to domain/windowing and clean architecture boundaries
Phase 1: Clean the boundaries and remove risky coupling

Created src/domain/windowing/ module:
- time.ts - clockToMinutes, minutesToClock
- policy.ts - windowRange, classifyWindowTiming, buildWindowDisplayPlan, getRunTimeContext
- fallback.ts - timeAwareBriefingFallback
- index.ts - public exports
- README.md - module documentation

Updated imports to use windowing module:
- src/domain/editorial/resolution/resolve-editorial.ts
- src/domain/editorial/resolution/validation.ts
- src/presenters/email/index.ts
- src/presenters/email/kit-advisory.ts
- src/presenters/email/next-day.ts
- src/presenters/email/time-aware.ts
- src/presenters/site/format-site.ts
- src/presenters/site/sections/hero.ts
- src/presenters/site/sections/window.ts
- src/presenters/site/sections/long-range.ts
- src/presenters/site/sections/daylight-utility.ts
- src/presenters/email/sections/long-range.ts

Added dependency rules documentation:
- docs/architecture.md - Added 'Dependency Rules' section
- src/domain/README.md - Added dependency rule and windowing reference

Architecture improvements:
- Domain no longer imports from presenters
- Window policy functions live in domain where they belong
- Presenters re-export from windowing for backwards compatibility
- Clear dependency direction: domain → lib, presenters → domain (for policy)

All tests pass, typecheck clean, artifacts regenerated.
